requirements.txt
tensorflow
numpy
opencv-python
matplotlib
scikit-learn
streamlit
Pillow
